Added authorization, books and transaction functional.

+++ src/AutoBookKeeper.Application/Services/UsersService.cs
@@ -1,3 +1,4 @@
+using AutoBookKeeper.Application.Exceptions;
 using AutoBookKeeper.Application.Interfaces;
 using AutoBookKeeper.Application.Mappers;
 using AutoBookKeeper.Application.Models;
@@ -25,19 +26,19 @@ public class UsersService : IUsersService
     public async Task<UserModel?> GetByIdAsync(Guid userId)
     {
         var user = await _usersRepository.GetByIdAsync(userId);
-        return ApplicationMapper.Mapper.Map<UserModel>(user);
+        return user == null ? null : ApplicationMapper.Mapper.Map<UserModel>(user);
     }
 
     public async Task<UserModel?> GetByUserNameAsync(string userName)
     {
-        var users = await _usersRepository.GetAsync(new UserSpecification(u => u.UserName == userName));
+        var users = await _usersRepository.GetAsync(UserSpecification.GetByName(userName));
         return ApplicationMapper.Mapper.Map<UserModel>(users.SingleOrDefault());
     }
 
     public async Task<UserModel?> GetByEmailAsync(string email)
     {
-        var users = await _usersRepository.GetAsync(new UserSpecification(u => u.Email == email));
-        return ApplicationMapper.Mapper.Map<UserModel>(users.FirstOrDefault());
+        var users = await _usersRepository.GetAsync(UserSpecification.GetByEmail(email));
+        return ApplicationMapper.Mapper.Map<UserModel>(users.SingleOrDefault());
     }
 
     public async Task<int> CountAsync()
@@ -77,35 +78,46 @@ public class UsersService : IUsersService
         
         var userEntity = ApplicationMapper.Mapper.Map<User>(user);
         userEntity.PasswordHash = passwordHash;
+
+        if ((await _usersRepository.GetAsync(UserSpecification.GetByName(user.UserName))).Any())
+            return AlreadyExistsResult("User with this name already exists");
         
-        // TODO check is this userName already exists
-        // TODO check is this email already exists
+        // todo check is this email already exists
         
         var result = await _usersRepository.CreateAsync(userEntity);
 
-        return ParseRepositoryResult(result);
+        return MappedRepositoryResult(result);
     }
 
     public async Task<OperationResult<UserModel>> UpdateAsync(UserModel user)
     {
         var userEntity = await _usersRepository.GetByIdAsync(user.Id);
-        if (userEntity == null) return ResultWhenUserNotFound();
+        if (userEntity == null) return NotFoundResult();
 
-        if (!string.IsNullOrWhiteSpace(user.UserName)) userEntity.UserName = user.UserName;
-        if (!string.IsNullOrWhiteSpace(user.Email)) userEntity.Email = user.Email;
-        
-        // TODO check is this userName already exists
-        // TODO check is this email already exists
+        if (!string.IsNullOrWhiteSpace(user.UserName))
+        {
+            userEntity.UserName = user.UserName;
+            
+            if ((await _usersRepository.GetAsync(UserSpecification.GetByName(user.UserName))).Any())
+                return AlreadyExistsResult("User with this name already exists");
+        }
+
+        if (!string.IsNullOrWhiteSpace(user.Email))
+        {
+            userEntity.Email = user.Email;
+            
+            // TODO check is new email already exists
+        }
         
         var result = await _usersRepository.UpdateAsync(userEntity);
         
-        return ParseRepositoryResult(result);
+        return MappedRepositoryResult(result);
     }
 
     public async Task<OperationResult<UserModel>> UpdateUserPassword(UserModel user, string currentPassword, string newPassword)
     {
         var userEntity = await _usersRepository.GetByIdAsync(user.Id);
-        if (userEntity == null) return ResultWhenUserNotFound();
+        if (userEntity == null) return NotFoundResult();
 
         if (!VerifyPassword(userEntity, currentPassword))
             return OperationResult<UserModel>.FromErrors(400, ["Invalid password"]);
@@ -114,20 +126,29 @@ public class UsersService : IUsersService
         
         var result = await _usersRepository.UpdateAsync(userEntity);
         
-        return ParseRepositoryResult(result);
+        return MappedRepositoryResult(result);
     }
 
     public async Task<OperationResult<UserModel>> DeleteAsync(UserModel user)
     {
-        var userEntity = ApplicationMapper.Mapper.Map<User>(user);
+        var userEntity = await _usersRepository.GetByIdAsync(user.Id);
+        if (userEntity == null) return NotFoundResult();
+        
         var result = await _usersRepository.DeleteAsync(userEntity);
 
-        return ParseRepositoryResult(result);
+        return MappedRepositoryResult(result);
     }
 
-    private static OperationResult<UserModel> ParseRepositoryResult(OperationResult<User> repositoryResult) => 
+    private static OperationResult<UserModel> MappedRepositoryResult(OperationResult<User> repositoryResult) => 
         repositoryResult.ToOperationResult(ApplicationMapper.Mapper.Map<UserModel>);
     
-    private static OperationResult<UserModel> ResultWhenUserNotFound() => 
-        OperationResult<UserModel>.FromErrors(404, ["User was not found"]);
+    private static OperationResult<UserModel> NotFoundResult(params string[] errors) => 
+        errors.Length > 0 ? 
+            new OperationResult<UserModel> {Status = 404, Exception = new NotFoundException(errors[0]), Errors = errors} : 
+            new OperationResult<UserModel> {Status = 404, Exception = new NotFoundException("User was not found"), Errors = ["User was not found"]};
+
+    private static OperationResult<UserModel> AlreadyExistsResult(params string[] errors) => 
+        errors.Length > 0 ? 
+            new OperationResult<UserModel> {Status = 400, Exception = new AlreadyExistsException(errors[0]), Errors = errors} : 
+            new OperationResult<UserModel> {Status = 400, Exception = new AlreadyExistsException("User already exists"), Errors = ["User already exists"]};
 }
